requirements.txt
python-telegram-bot==21.3
pandas==2.2.3
pandas_ta==0.3.14b0
ta==0.11.0
requests==2.32.3
httpx==0.24.1
nest_asyncio==1.6.0

